Your commit message here

+++ routes/todo.js
@@ -6,8 +6,7 @@ router.get('/', async function (req, res) {
   try {
     const tasks = await Task.find({});
     return res.render('todo', {
-      title: "Home",
-      task: tasks,
+      task: tasks
     });
   } catch (err) {
     console.error('Error in fetching tasks from db', err);
@@ -31,20 +30,25 @@ router.post('/create-task', async function (req, res) {
   }
 });
 
-
 router.get('/delete-task', async function (req, res) {
   try {
     const ids = Array.isArray(req.query.id) ? req.query.id : [req.query.id];
+    const action = req.query.action; 
 
     await Promise.all(ids.map(async (taskId) => {
-      await Task.findByIdAndDelete(taskId);
+      if (action === "DELETE TASK") {
+        await Task.findByIdAndDelete(taskId);
+      } else if (action === "Mark As Completed") {
+        await Task.findByIdAndUpdate(taskId, { status: 'completed' });
+      }
     }));
 
     return res.redirect('back');
   } catch (err) {
-    console.error('Error in deleting tasks', err);
+    console.error('Error in managing tasks', err);
     res.status(500).send('Internal Server Error');
   }
 });
+
 
 module.exports=router;
